fix name collisions on inherited variables
On function contexts, inherited instance variables and instance
variables were not checked for matches, so functions could end up
requiring the same parameter twice.

This commit also reintroduces an intermermediate this substring on the
variable names to avoid inherited variable names colliding with locals
and globals.

Fixing this also showed a bug on the function namespace parameter
substitution which is fixed now.