reimplement function namespace parameters
The previous 2 pass implementation couldn't had problems on function
scopes, as it found the effects of the first pass and bailed out because
it thought it wasn't able to do a replacement.

This commit does it on a single more complex pass.

To know exactly what was happening, try the updated
"jsfx-samples/namespace_parameter" on the previous commit.