Don't add "this" to function variables names.
This commit hopefully fixes either variable name collision in functions
with autogenerated instance variables or duplicated generated variables
because of mismatches prefixes.

"this" is not used anymore. Just a dollar '$' for brevity.

"_make_key" is modified to remove consecutive dollar signs.

The logic of adding prefixes is now both done on variable classification
and in function call resolutions. It behaves well on all the tests.